bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
yt-dlp>=2024.4.9
PyNaCl>=1.5.0
flask>=2.3.2